requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.1.1
requests==2.31.0
PyYAML==6.0.1
Flask==3.0.3
flask-cors==4.0.1
crypto==1.4.1
psycopg==3.1.19
python-telegram-bot[job-queue]
gunicorn
